+++ src/Components/Movies/Movies.tsx
@@ -6,31 +6,24 @@ import { useNavigate } from "react-router-dom";
 import ShowMovies from "./ShowMovies";
 import Filter from "../Filter";
 import { RouteKey } from "../../navigation/router";
-
-export interface TMovie {
-  Title: string;
-  Year: string;
-  imdbID: string;
-  Type: string;
-  Poster: string;
-}
+import { TMovie } from "../../api";
+import { API_URL } from "../../api/shared/constants";
 
 const Movies = () => {
-
+  
   const [movies, setMovies] = useState<TMovie[]>([]);
   const [filteredMovies, setFilteredMovies] = useState<TMovie[]>([]);
   const [years, setYears] = useState<string[]>([]);
-
+  
   const [user] = useAuthState(auth);
   const navigate = useNavigate();
-
 
   useEffect(() => {
     if (!user) navigate(RouteKey.Index);
   }, [user]);
 
   const getResponse = async () => {
-    await axios.get(`https://www.omdbapi.com/?s=Batman&apikey=887f9f42`).then(res => {
+    await axios.get(API_URL).then(res => {
       const data: TMovie[] = res.data.Search;
       setMovies([...data]);
       setYears(data.map(movie => movie.Year).sort((a, b) => a.localeCompare(b)));
