+++ src/api/shared/constants.ts
@@ -1,0 +1,1 @@
+export const API_URL = "https://www.omdbapi.com/?s=Batman&apikey=887f9f42";
