+++ src/api/types.ts
@@ -1,0 +1,7 @@
+export type TMovie = {
+    Title: string;
+    Year: string;
+    imdbID: string;
+    Type: string;
+    Poster: string;
+}
